requirements.txt
fastmcp
garminconnect==0.2.28
python-dotenv==1.1.1
fastapi